Property
In property the defined methods to create are temporary. A setType method is required to make sure the type is within the types that we have. On the UI standpoint it can be represented as a popup menu or some other type of menu to make sure they don't put something that doesn't exist